Separate option-related functions to options.cpp
This also necessitated adjusting the cmake CMakeLists.txt file. It should be noted here that I am using a GLOB to collect all files in the src/ directory and compile them at once. CMake itself does not recommend this but offers no suitable alternative. The only feasible alternative would be to specify each source file individually, which just creates a lot of unnecessary and extremely inflexible boilerplate code. Perhaps I misunderstand what the point of that kind of structure is, but for this comparitively small project, it would be overkill, I believe.