Remove automated time tracking including win hooks
This also changes the mission of dlog. It no longer aims to provide
automated time tracking capabilities but is now completely manual. This
is because, in my recent experiences over the past few months, automated
time tracking is rarely worth it. While it can be quite convenient to
track the time spent in various programs, consolidating all that data
into a meaningful form is often quite difficult and/or impossible, not
to mention that it makes tracking time spent outside of programs (for
instance when walking away from the computer) likewise impossible.

Instead of maintaining two different ways to track time spent (manual
and automatic), it was decided to focus on one (manual) and remove the
other entirely, at least for the moment.