Remove formatting of recent times as "now"
This change was decided because in the commands where "now" was
outputted, it provided no benefit whatsoever. It was outputted in the
commands "dlog start" and "dlog fill", but in both cases the output
would *always* be "now", thus no information was actually provided.

Returning to normal local time strings also allows a user to review
their recent fill commands by scrolling up through their terminal
output. If all commands outputted "now", the information of when the
command was run would be lost.