Fix summary tests ignoring temporal flags (-a etc)
Cause: the internal logic of the filtering algorithm checks for the
presence of the from and to values on the summary instance. If those
exist, it no longer checks the -a, -y, -w, and -d flags. This is fine
for the actual runtime since trying to enter -f/-t together with the
flags results in a conflicting argument clap error anyway, but it did
break the tests that used the -a flag.